+++ package.json
@@ -48,6 +48,9 @@
     "electron-squirrel-startup": "^1.0.0",
     "estree-walker": "^1.0.1",
     "monaco-editor": "^0.23.0",
+    "node-pty": "^0.10.1",
+    "xterm": "^4.12.0",
+    "xterm-addon-fit": "^0.5.0",
     "yootils": "^0.2.0"
   },
   "devDependencies": {
