+++ src/index.js
@@ -2,11 +2,10 @@ const { app, BrowserWindow, dialog, ipcMain } = require('electron');
 const createApplicationMenu = require('./application-menu');
 const path = require('path');
 const fs = require('fs')
+const os = require('os');
+const pty = require('node-pty');
 
-require('electron-reload')(__dirname, {
-  electron: path.join(__dirname, '../node_modules', '.bin', 'electron'),
-  awaitWriteFinish: true
-});
+
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) {
@@ -83,6 +82,28 @@ const createWindow = exports.createWindow = () => {
     newWindow = null;
   });
 
+  var shell = os.platform() === "win32" ? "powershell.exe" : "bash";
+  var ptyProcess = pty.spawn(shell, [], {
+          name: 'xterm-color',
+          cols: 80,
+          rows: 24,
+          cwd: process.env.HOME,
+          env: process.env
+      });
+    
+    ptyProcess.on("data", (data) => {
+      newWindow.webContents.send("terminal-incData", data);
+    });
+
+    ipcMain.on("terminal-into", (event, data)=> {
+      ptyProcess.write(data);
+    })
+
+require('electron-reload')(__dirname, {
+  electron: path.join(__dirname, '../node_modules', '.bin', 'electron'),
+  awaitWriteFinish: true
+});
+
   windows.add(newWindow);
   return newWindow;
 };
@@ -91,7 +112,7 @@ const getFileFromUser = exports.getFileFromUser = async (targetWindow) => {
   const files = await dialog.showOpenDialog(targetWindow, {
     properties: ['openFile'],
   });
-  
+
   if(files) {
     if (files) { openFile(targetWindow, files.filePaths[0]); }
   }
