--- src/renderer.js
@@ -1,5 +1,0 @@
-const {ipcRenderer} = require('electron');
-
-export default ipcRenderer.on('file-opened', function (evt, message) {
-    console.log(message); // Returns: {'SAVED': 'File Saved'}
-});
